pyproject.toml
[project]
name = "videox-fun"
version = "1.0.1"
description = "VideoX-Fun is a video generation pipeline that can be used to generate AI images and videos, as well as to train baseline and Lora models for Diffusion Transformer. We support direct prediction from pre-trained baseline models to generate videos with different resolutions, durations, and FPS. Additionally, we also support users in training their own baseline and Lora models to perform specific style transformations."
license = { file = "LICENSE" }
dependencies = [
    "Pillow",
    "einops",
    "safetensors",
    "timm",
    "tomesd",
    "torch>=2.1.2",
    "torchdiffeq",
    "torchsde",
    "decord",
    "datasets",
    "numpy",
    "scikit-image",
    "opencv-python",
    "omegaconf",
    "SentencePiece",
    "albumentations",
    "imageio[ffmpeg]",
    "imageio[pyav]",
    "tensorboard",
    "beautifulsoup4",
    "ftfy",
    "func_timeout",
    "accelerate>=0.25.0",
    "gradio>=3.41.2",
    "diffusers>=0.30.1",
    "transformers>=4.46.2",
]

[project.urls]
Repository = "https://github.com/aigc-apps/VideoX-Fun"
# Used by Comfy Registry https://comfyregistry.org

[tool.setuptools]
packages = ["videox_fun"]

[tool.comfy]
PublisherId = "bubbliiiing"
DisplayName = "VideoX-Fun"
Icon = ""
